Remove FixedLengthBinary and VariableLengthBinary traits
### What
Remove FixedLengthBinary and VariableLengthBinary traits. Also impl `Binary` `Borrow`, `AsRef`, and `AsMut` for `ArrayBinary<N>`.

### Why
They add complexity to the concepts a developer needs to learn when interacting with binary that is not outsized by the benefits the type add.

The types make it possible for functions to indicate if when accepting a mutable reference to a binary if that function is going to modify the binary in a way that will change its size or not. In conversations with @jonjove he mentioned he has only found a use for this benefit in tests and not really in contract code.

I think we can get by with the two types, Binary and ArrayBinary. I'm aware of these use cases:
- Variable/undefined length binaries can be passed as `&Binary` / `&mut Binary` (values and size may be changed).
- Fixed and specified length binaries can be passed as `&ArrayBinary<_>` / `&mut ArrayBinary<_>` (values may be changed).
- Fixed and unspecified length binaries can be passed as `&ArrayBinary<N>` / `&mut ArrayBinary<N>` (values may be changed, size can be unspecified by use of a const generic).

The one use case the concrete types don't satisfy is if you want to accept a fixed size array mutable so you can't change size, and you don't know at compile time what the size will be. In this case you'd have to use a `dyn ArrayBinary<N>` I guess, I'm not even sure if that would work with the const generic. I think this use case is pretty narrow and unlikely to be a huge problem.

For any function wanting to accept either a `Binary` or `ArrayBinary` the traits are unnecessary, because the function can accept either by accepting `Borrow<Binary>` / `BorrowMut<Binary>`, or `AsRef<Binary>` / `AsMut<Binary>`.