Remove panic-catching and fix tests that use newly-working native try_call (#746)
fixes https://github.com/stellar/rs-soroban-env/issues/462
fixes https://github.com/stellar/rs-soroban-sdk/issues/642

This "fixes" the above bugs by removing the panic-catching machinery
from the SDK as discussed with @leighmcculloch -- the host now does a
`catch_panic` inside `try_call` and calls `rollback` itself, so it
functions "like the VM". This change is mostly just removing stuff from
the SDK and tweaking tests to expect correct behaviour.

I think the only slightly hinky thing in here is that the tests assert
on a _specific_ status that differs depending on whether the panic was
caught natively by the host or due to a vm trap. I think this is the
best we can do here and now, but we might want to make a note of it
somewhere as a consideration when assigning the final error-code
repertoire: it might be better to merge these two error codes and just
restrict the difference between a failed VM `try_call` and a failed
native `try_call` to differences in `DebugEvent`s

cc @jayz22

Co-authored-by: Leigh McCulloch <351529+leighmcculloch@users.noreply.github.com>