Remove type checks for values delivered from the host (#751)
### What
Remove type checks for values delivered by the host to the guest.

### Why
We said a while ago that values coming from the host to the guest would
be trusted. We make this assumption in many places already. In various
we still do a type check on the guest side, which seems to be
inconsistent and unnecessary. This removes unnecessary branches from
contracts and reduces their size.

### Merging
This is intended to be merged to `main` after #750.