Add Env::as_contract for testutils (#761)
### What
Add Env::as_contract that executes a function inside a test frame
artificially created that may not reference a contract that is actually
deployed.

### Why
To support providing functionality where logic is executed as a
contract, without the contract having to provide that logic itself. This
is helpful in testing.

Dependent on https://github.com/stellar/rs-soroban-env/pull/565

Close #739