Fix tests when run in some configurations (#775)
### What
Fix tests when run in some configurations where testutils isn't
activated.

### Why
We should always configure code intended for tests with `test` and
`testutils`, not just `testutils`. I made this mistake in a recent PR.
It breaks running tests in some IDEs, although by happenstance this
doesn't show up in CI because of feature unification that happens during
the CI builds.