(minimal) Adapt to introduction of Env::Error (#837)
This is a "minimal" version of
https://github.com/stellar/rs-soroban-sdk/pull/833, accompanying /
adapting to https://github.com/stellar/rs-soroban-env/pull/638.

If this minimal approach is preferred, this PR and
https://github.com/stellar/rs-soroban-env/pull/638 should be merged
together. Otherwise the more-invasive pair #833 and
https://github.com/stellar/rs-soroban-env/pull/633 should be merged
together.

This one does much less rearranging and makes no difference to the
emitted code at all. It just pushes around where the non-infallible
`Env::Error` cases get escalated: formerly it happened in `impl Env for
CheckedEnv`, now there's a new internal SDK function
`internal::reject_err` that does the work. Infallible errors also need
to be explicitly unwrapped in a bunch of places but this is just a
type-level thing, operationally it's a no-op.

(I should also note: I could do an "even more minimal" version of this
where `soroban_sdk::env::Env` _does not implement_
`soroban_env_common:env::Env`, in which case the "error rejections"
could be buried in the macro-generated code in the SDK and we could
remove all the `.unwrap_infallible` calls in the SDK bodies. I don't
know how important it is for the SDK's Env to implement the common Env.)