Allow same named types and functions (#839)
### What
Change the names of the static variables that contain the spec XDR to
include a type or function qualifier for the kind of entry the spec
contains.

### Why
To allow same named types and functions in contracts. When a type and
function have the same name today they result in two static variables
that have the same names. Adding a qualifier to what they are to those
variables names will stop them from colliding.

Close #828