Make soroban_token_spec::spec_xdr easier to keep updated (#845)
### What
Change soroban_token_spec::spec_xdr to generate the final value from
constant inputs, and add a test that checks that the constant inputs are
valid.

### Why
The spec_xdr function has to know ahead of time what the final size of
the byte slice will be because the function is a const function.
Additionally the panics in the const function cannot construct an error
message containing the expected length. Whenever we change the format of
the spec xdr it's difficult to know what the new length is. We get a
very opaque error message saying the length is wrong but not what to set
it to.

By moving the inputs and expected length to be constants, and add a test
that asserts the length matches the sum of the inputs, we provide a way
to get the length for updating it.