Add a utility to call `__check_auth` in tests. (#887)
### What

Add a utility to call `__check_auth` in tests. 

### Why

This probably could be achieved differently by just lifting the
restriction on `__check_auth` callers in tests. However, I think that
the explicit approach better conveys the fact that there is no way to
call `__check_auth` on-chain and also adds a bit more type safety for
the arguments.

### Known limitations

N/A