Hide Env.host() in docs (#922)
### What
Hide Env.host() in docs.

### Why
The `Env` `host` function returns a `Host`, which is an internal type
that is not documented. It isn't intended for routine use and we haven't
really made a plan for how power users or edge cases might use it. At
the moment it is just noise in the docs, somewhat distracting and
ambiguous especially as it has no docs. For these types of functions
that we don't really expect people to use right now we typically hide
them.