Remove dependence on auth when registering stellar asset contract in tests (#942)
### What
Remove dependence on auth when registering stellar asset contract in
tests.

### Why
Registering a stellar asset contract in tests involves creating the
token, and then modifying the admin. The modification of the admin is
currently done by calling the set_admin function on the token, which
requires authorization be setup for the issuer to approve, or for
recording/mocking to be enabled. The problem with this is that calling
the register function will squash any existing authorization setup in
the environment. Ideally we would have a way to snapshot the current
authorization, set issuer auth, make the call, then reset it back to
what it was. Unfortunately that's not yet implemented, but is being
tracked with https://github.com/stellar/rs-soroban-env/issues/785.

A simple work around is for us to setup the ledger entry manually. In
some perspectives this might be a hack given that the token contract has
a function for setting the admin. However, we already do ledger entry
setup for registering stellar asset contracts, so the impact seems net
neutral.