Remove Address::account_id facility (#946)
### What
Remove Address::account_id and Address::from_account_id facilities.

### Why
The utility of these methods is pretty confusing, and not really aligned
with where we're headed for a couple reasons.

The `account_id` function returns the ed25519 public key of an Account
ID. This is the first place in the Stellar protocol where we expose an
Account IDs key without context for what type of key it is. Account IDs
are usually transmitted with a discriminant so that we can add new key
types in the future without applications needing to be redeveloped. We
could address this by renaming the function to
`account_id_ed25519_public_key`.

However, renaming the function to that name makes it even clearer that
there's a bigger issue at play. The only thing someone might do with an
ed25519 public key is to use it for identification or signature
verification. For identification contracts should use the `Address` type
directly.

That leaves the only use case as signature verification of an Account
ID's public key. However, signature verification of a master key is
dangerous. It makes the assumption that it is the signer of the account,
and contracts have no access to signers of accounts to know if that is
the case. Any contract that verifies signatures for an Account ID is
ignoring the multisig signers setup of an account and could be verifying
a key that is not a signer of the account.

While it feels natural to include these functions, I think they're
harmful. We need to steer people towards only using `Address`. Outside
of contracts Stellar developers use strkeys whenever interacting with
IDs on the Stellar network. `Address` is the strkey of contracts, and
folks need to use that.

Note that if there are use cases in tests where account ID `Address`es
need to be tested, this can still be achieved by creating an
`xdr::ScAddress` containing an `xdr::AccountId`, and then converting
that to an `Address`.

Related conversation is at
https://discord.com/channels/897514728459468821/1108510759513763860