Move AuthorizationContext into SDK and add address method (#951)
### What
Move AuthorizationContext into the SDK, renamed to AuthContext, and add
method `AuthContext::contract()` that returns the `Address` for the
contract in the context.

### Why
I removed `BytesN<32>` everywhere I could for contract IDs, but
AuthorizationContext still has to contain a BytesN<32> because the
environment generates the value and the environment hasn't been updated
yet to use Address everywhere.

I had changed `Address` to disallow creating it from `BytesN<32>` so
that we have developers go all in on `Address`. To have the
`AuthorizationContext` give the user an `Address` it needs to be in the
SDK so it can call the private `Address::from_contract_id` function.

I renamed the type because we use the term `Auth` in docs and most
places, so the long form seems unnecessary.

Moving the type is probably a good long term move given that auth is a
core concept, there's little need to separate auth logic into a separate
crate now. I left the auth test utility in the separate crate to address
separately.