Rename Client::contract to Client::address (#954)
### What
Rename the `contract` field of generated clients to `address`.

### Why
I recently renamed the `contract_id` field of generated clients to
`contract` when I changed their type from a contract ID in a
`BytesN<32>` to an `Address`.

I just went through and updated all the examples to use the `contract`
field, and what I discovered is the code is now pretty ambiguous and
hard to read. In plenty of places we have things like `token.contract`
or `contract.contract`, and it's really not clear what I have a
reference to. Is it a contract client, or a contract wasm, or something
else?

What I also noticed is that at some point we'd added an `address()`
function to generated clients, and code that was using that function
read really clearly.

Renaming the field to `address` to get that same clarity.