Move json logic from soroban-spec to soroban-spec-json (#972)
### What
Move json logic from soroban-spec to soroban-spec-json.

### Why
As we write more code generators the soroban-spec crate is going to
become a big soup of logic with lots of dependencies.

We already see that in some ways with the json generator living in
soroban-spec, which is another point that the serde and serde_json, both
large crates, get pulled into soroban-sdk macro builds. We can also see
that in the v0.8.5 patch release which contained the typescript
generator.

These different generators will have their own sets of dependencies, and
their own owners/maintainers/reviewers. The only generator that is owned
by the folks who maintain this repository is the rust generator, and it
should continue to live here, but other generators can live elsewhere.

The json generator is only used by the soroban-cli and not the
soroban-sdk, so we may wish to move the crate into the soroban-tools
repo as well, although we can discuss that independent of this change.

Using a pattern for organizing the code generators that allows us to
keep the code separate, will make it easier for us to place them in
places that the maintainers for those code generators live.

For https://github.com/stellar/rs-soroban-sdk/issues/968