Remove Set (#979)
### What
Remove the Set type.

### Why
As part of #379 we are changing the interfaces of all our collection
types. Set is a collection type that might be useful, but it is doc
hidden right now, and not in use. There's no evidence yet we need it.

Instead of spending time rewriting its interface to match that of Vec
and Map, we will remove it, and then bring it back when the time is
right.

Close #379