Change Vec iter to not create many Vecs (#999)
### What
Change Vec iter to not create a new Vec on each step of the iteration.

### Why

The Vec iterator currently makes a host copy of the vec on every step of
the iteration. This is because it uses the slice function to get a new
vec with the first or last item removed. Even though it results in
optimized guest code because the iterator is only a single u64 handle,
those optimizations are outsized by the calls to the host that
repeatedly duplicate the vec.

Close #189

### Alternatives

There might be some advantage in having the slice function not copy the
vec on the host side. The host could lazily reuse the vec backing array
for read, pop, and slice operations. I don't think the guest savings
would be worth the host complexity though. Additionally every step of an
iteration would get two host fn calls instead of one, so in the end it
is difficult to know if it would actually be better.

### Perf Diff

#### Code

```rust
let env = Env::default();

let vec = vec![&env, 0, 1, 2, 3, 4];

let mut iter = vec.iter();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();
_ = iter.next();

env.budget().print()
```

#### Code Size

```diff
@@ -1 +1 @@
--rwxr-xr-x  1 leighmcculloch  staff  1210 Jun 20 22:25 test_iter.wasm
+-rwxr-xr-x  1 leighmcculloch  staff  934 Jun 20 22:30 test_iter.wasm
```

#### Budget

```diff
@@ -1,15 +1,15 @@
 =======================================================
-Cpu limit: 40000000; used: 35996
-Mem limit: 52428800; used: 504
+Cpu limit: 40000000; used: 6835
+Mem limit: 52428800; used: 104
 =======================================================
 CostType                 cpu_insns      mem_bytes      
 WasmInsnExec             0              0              
 WasmMemAlloc             0              0              
-HostMemAlloc             28200          504            
-HostMemCpy               276            0              
+HostMemAlloc             4700           104            
+HostMemCpy               46             0              
 HostMemCmp               0              0              
 InvokeHostFunction       0              0              
-VisitObject              475            0              
+VisitObject              114            0              
 ValXdrConv               0              0              
 ValSer                   0              0              
 ValDeser                 0              0              
@@ -23,7 +23,7 @@ VmMemRead                0              0
 VmMemWrite               0              0              
 VmInstantiation          0              0              
 InvokeVmFunction         0              0              
-ChargeBudget             7020           0              
+ChargeBudget             1950           0              
 ComputeKeccak256Hash     0              0              
 ComputeEcdsaSecp256k1Key 0              0              
 ComputeEcdsaSecp256k1Sig 0              0              
```