Mark exported fns as extern (#1002)
### What
Mark exported fns as `extern`, which by default causes them to be marked
as `extern "C"`.

### Why
Technically the Rust ABI is an internal detail. For an ABI stability
guarantee I think we need to use the C ABI, which we can do by
specifying `extern`.

A quick test with test vectors in this repo show that the resulting
binary is identical with and without extern, so it really doesn't seems
like it is need today.

Still adding because technically it seems like the correct thing to do
even if it has no material impact today.