Rename logger to logs (#1005)
### What
Rename env.logger() to env.logs().

### Why
The logger function is inconsistently named compared to the other
functions for accessing resources. E.g. `events`, `storage`, `ledger`. I
keep looking for logs and having to remember it is logger. The type is
used not just for the act of logging, but also the act accessing logs in
tests.

Note that the deploy functionality is under `deployer`, but it's not one
that as easily renamed to `deploy` or `deploys`.

The rename is done backwards compatible with a deprecated notice, so
folks will get a meaningful warning message and do not need to go
hunting for the new function.