Add #[contract] macro (#1006)
### What
Add `#[contract]` macro that generates a contract client for the
contract, and make #[contractimpl] only add functions to the existing
contract client.

### Why
To support multiple `#[contractimpl]` per contract.

For contracts that have multiple impls, it is not ideal if there are
multiple clients for that one contract. Ideally clients appear in the
same way everywhere, and anyone importing the contract will see a single
flat client.

Since macros cannot coordinate, there's no way for the first
contractimpl to output the client struct and others to skip it. Today
what happens is a compile error is shown because there are multiple
clients generated.

Unblocks:
- #952 
- #965 
- And general better interoperability through trait sharing.

Close #268