Add some auth test contracts (#938)
### What
Add some auth test contracts that exercise different auth scenarios.

### Why
I initially started writing these as a learning experience, but would
like to commit them as a way to capture existing behavior that contract
developers are exposed to and therefore relying on. As auth evolves in
the coming releases, changes to these tests will make it clearer to me
and others working at the application layer on how auth is changing.