Split up the token interface trait (#1012)
### What
Separate admin functions in the token interface out to a separate trait.

### Why
The primary standard interface of tokens won't include the admin
functions. Although many tokens on Soroban will have the admin functions
because all Stellar Asset Contracts will have them.

We still want it to be convenient for folks to follow that same admin
pattern if its a fit, but contracts shouldn't have to implement them if
their follow a different pattern for admin operations.

This issue was discussed [on
Discord](https://discord.com/channels/897514728459468821/1114325793926037605),
and some folks suggested allowances should also be broken into a
separate trait. However this seemed more controversial and there were
also opinions that they shouldn't be separated, so I've kept them in for
now.

Note that the Spec value in the SDK will continue to hold the entire
token interface. It's not particularly useful to split that up, as the
only user of it is the soroban-cli when interacting with Stellar Asset
Contracts. I've hidden it from docs since it isn't routinely useful to
most.

Close #965