Bring back the symbol macro (#1019)
### What
Bring back the `symbol!` macro from the past, in the form of
`symbol_short!`.

### Why
The symbol! macro existed to ensure that when short symbols were
generated, they were entirely generated at compile time. The goal was
that the string and the conversion code would never end up in the file
wasm.

At some point we removed the macro and replaced it with a const
function. However const functions are only const evaluated at compile
time if they are called in a const context.

The impact is that anytime a short symbol is used without the macro it
increases the size of the wasm by about 200 bytes, because both the
original string and the conversion code are embedded into the wasm.

Close #927

### Perf Diff

```diff
--rwxr-xr-x  1 leighmcculloch  staff  858 Jun 23 20:33 test_errors.wasm
+-rwxr-xr-x  1 leighmcculloch  staff  658 Jun 23 20:36 test_errors.wasm
--rwxr-xr-x  1 leighmcculloch  staff  954 Jun 23 20:32 test_events.wasm
+-rwxr-xr-x  1 leighmcculloch  staff  667 Jun 23 20:36 test_events.wasm
--rwxr-xr-x  1 leighmcculloch  staff  1025 Jun 23 20:32 test_invoke_contract.wasm
+-rwxr-xr-x  1 leighmcculloch  staff  843 Jun 23 20:35 test_invoke_contract.wasm
```