Add test vector for workspace setups where contract types live in a lib (#1070)
### What
Add test vector for workspace setups where contract types live in a lib.

### Why
In https://github.com/stellar/rs-soroban-sdk/issues/1063 it was
mentioned that folks were having trouble with using types in other libs.
Added this test vector to confirm if that was an issue or not. Appears
to work fine, and this is a use case that is important, so committing
the test vector.