Add Vec::to_vals (#1087)
### What
Add Vec::to_vals function that converts a Vec<T> to a Vec<Val>.

### Why
It's reasonably common, at least in tests and internally, that we need
to construct Vec's of Val's. There are different ways to do this by
constructing a Vec of Val::from calls, to using tuples with always
convert to Vec<Val>. We should have a function to help us do it.

I hope to use the function in #1023.