docs: Fixing a broken link in the SDK's `lib.rs` file (#1099)
### What

A link in the `lib.rs` file of the SDK previously directed users to the
`how-to-guides` category on the website. I've changed that link to lead
to the `basic-tutorials` page, and have added a link to the
`advanced-tutorials` page.

### Why

The `how-to-guides` page no longer exists on the stellar/soroban-docs
repo. We're seeing some 404 errors originating from `docs.rs` URLs.

### Known limitations

N/A