Use deterministic randomness in arbitrary tests (#1120)
### What

Use a seeded StdRng instead of thread_rng in arbitrary tests.

### Why

Per https://github.com/stellar/rs-soroban-env/issues/810 randomized test
cases should be deterministic.

### Known limitations

This is different from the approach in soroban-env-host from
https://github.com/stellar/rs-soroban-env/pull/1124, where a test prng
is attached to the env / host.

There are two more uses of thread_rng in this crate, but I don't know
enough about that code to know how they should be removed, i.e. if they
should follow a pattern similar to soroban-env-host. I am willing to
replace all uses of thread_rng if given guidance about the preferred way
to do it.

The implementation of StdRng is allowed to change in the future, so it's
possible the exact numbers generated could change. I could import
rand_chacha instead if desired.

Co-authored-by: Leigh McCulloch <351529+leighmcculloch@users.noreply.github.com>