Fix docs about prng seed (#1128)
### What
Remove the statement from the prng seed function docs that states that
the base and local seed are combined.

### Why
They are not combined. Calling seed does what it sounds like it would do
with any prng, it sets a new seed replacing the base seed set.

Possibly we should stop using the term "base seed" as it makes it sound
like there's some relationship. But at the least we should remove this
incorrect comment about seed interaction.