Remove use of rand from generated addresses, nonces, salts, and issuer pks (#1135)
### What
Remove use of rand from generated addresses, nonces, salts, and issuer
pks, replace with sequential values instead.

### Why
The SDK generates random addresses, nonces, salts, and issuer pks. This
randomness makes it near impossible to write any test that produces
predictable and consistent total system state.

The SDK doesn't need to be using random values for these things. It just
needs to produce values that don't matter and that are unique.

Predictable and consistent values make it easier to write tests that can
be used over time to assess the consistency of a contract. This is an
important attribute of tests so that folks can evaluate if their
contract is consistent even with dependency upgrades, SDK upgrades, and
protocol upgrades.

This change also introduces the concept of a `Snapshot`, which is a
superset of the existing `LedgerSnapshot`. The `LedgerSnapshot` contains
everything needed to load in a ledger from a network or other source.
`Snapshot` contains everything needed to rehydrate the state of an SDK
testutils backed Env.

Related discussion in Discord:
https://discord.com/channels/897514728459468821/1171226885171191949

For #1082