Replace String::from_slice with from_str (#1139)
### What
Replace String::from_slice with String::from_str.

### Why
Most of the time functions that convert from a str into something else
are named `from_str`. That's usualy because folks are implementing the
FromStr trait. We're not implementing it here because it isn't entirely
appropriate. FromStr is only for fallible conversions and this is an
infallible conversion. But it still feels right to use the 'str' naming
here instead of 'slice'. Especially since it is uncommon to talk about
strs as slices in rust, even though strs can be sliced.

For #1138