Rename token::Interface/Client to TokenInterface/Client (#1149)
### What
Rename token::Interface/Client to TokenInterface/Client.

### Why
When I've used the token::Interface it ends up with less than readable
code because I end up with this generic "Interface" identifier in my
code. If other interfaces follow suit it gets even worse when
implementing multiple interfaces.

The rename happens in a backwards compatible way. The types are
reexported as their old names, and the old names are hidden from docs,
and have deprecated notices. The deprecated notice informs a developer
what they need to change to update to the change. The hidden doc is
helpful for the upcoming release that will be a first release and would
be odd to display deprecated fns in the docs given any fns were release
candidates anyway and the deprecated notice is a courtesy and
convenience.