Add auth to Env snapshot (#1158)
### What
Include auth details in the Env snapshot that is persisted to file at
the end of tests.

### Why
The details about what auth is performed during a test run is not
currently captured in the Env snapshot, but it is something that is an
important behavior of a test that should be tracked to test for
consistency over time.

Close #1152 

Dependent on:
- https://github.com/stellar/rs-soroban-env/pull/1214
- https://github.com/stellar/rs-soroban-sdk/pull/1157
- https://github.com/stellar/rs-soroban-sdk/pull/1153